Add files via upload
Basically I found the average karma for every post in a given subreddit while looking at average account age. 

In this I found that r/depression actually has a relatively low average reddit karma score compared to other subreddits however one important thing to note is that the average account age is also a lot younger than other similar sized subreddits. 

Therefore our conclusion is that it is hard to determine if reddit karma is a good indication of bad mental health because from the data it looks like a good chunk of people creates a brand new account to make a post in subreddits like r/depression